requirements.txt
openai-whisper==20231117
yt-dlp>=2024.1.0
torch>=2.0.0
torchaudio>=2.0.0
youtube-transcript-api>=1.6.0 